// frontend/src/apisv2beta1/run/models/ArtifactArtifactType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * backend/api/v2beta1/run.proto
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: version not set
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *  - TYPE_UNSPECIFIED: default; treated as "not set"
 * reject if unset.
 * @export
 */
export const ArtifactArtifactType = {
  TYPE_UNSPECIFIED: 'TYPE_UNSPECIFIED',
  Artifact: 'Artifact',
  Model: 'Model',
  Dataset: 'Dataset',
  HTML: 'HTML',
  Markdown: 'Markdown',
  Metric: 'Metric',
  ClassificationMetric: 'ClassificationMetric',
  SlicedClassificationMetric: 'SlicedClassificationMetric',
} as const;
export type ArtifactArtifactType = (typeof ArtifactArtifactType)[keyof typeof ArtifactArtifactType];

export function instanceOfArtifactArtifactType(value: any): boolean {
  for (const key in ArtifactArtifactType) {
    if (Object.prototype.hasOwnProperty.call(ArtifactArtifactType, key)) {
      if (ArtifactArtifactType[key as keyof typeof ArtifactArtifactType] === value) {
        return true;
      }
    }
  }
  return false;
}

export function ArtifactArtifactTypeFromJSON(json: any): ArtifactArtifactType {
  return ArtifactArtifactTypeFromJSONTyped(json, false);
}

export function ArtifactArtifactTypeFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): ArtifactArtifactType {
  return json as ArtifactArtifactType;
}

export function ArtifactArtifactTypeToJSON(value?: ArtifactArtifactType | null): any {
  return value as any;
}

export function ArtifactArtifactTypeToJSONTyped(
  value: any,
  ignoreDiscriminator: boolean,
): ArtifactArtifactType {
  return value as ArtifactArtifactType;
}