// frontend/src/apisv2beta1/run/models/InputOutputsIOParameter.ts
/* tslint:disable */
/* eslint-disable */
/**
 * backend/api/v2beta1/run.proto
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: version not set
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { V2beta1IOType } from './V2beta1IOType';
import {
  V2beta1IOTypeFromJSON,
  V2beta1IOTypeFromJSONTyped,
  V2beta1IOTypeToJSON,
  V2beta1IOTypeToJSONTyped,
} from './V2beta1IOType';
import type { V2beta1IOProducer } from './V2beta1IOProducer';
import {
  V2beta1IOProducerFromJSON,
  V2beta1IOProducerFromJSONTyped,
  V2beta1IOProducerToJSON,
  V2beta1IOProducerToJSONTyped,
} from './V2beta1IOProducer';

/**
 *
 * @export
 * @interface InputOutputsIOParameter
 */
export interface InputOutputsIOParameter {
  /**
   *
   * @type {object}
   * @memberof InputOutputsIOParameter
   */
  value?: object;
  /**
   *
   * @type {V2beta1IOType}
   * @memberof InputOutputsIOParameter
   */
  type?: V2beta1IOType;
  /**
   *
   * @type {string}
   * @memberof InputOutputsIOParameter
   */
  parameter_key?: string;
  /**
   *
   * @type {V2beta1IOProducer}
   * @memberof InputOutputsIOParameter
   */
  producer?: V2beta1IOProducer;
}

/**
 * Check if a given object implements the InputOutputsIOParameter interface.
 */
export function instanceOfInputOutputsIOParameter(value: object): value is InputOutputsIOParameter {
  return true;
}

export function InputOutputsIOParameterFromJSON(json: any): InputOutputsIOParameter {
  return InputOutputsIOParameterFromJSONTyped(json, false);
}

export function InputOutputsIOParameterFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): InputOutputsIOParameter {
  if (json == null) {
    return json;
  }
  return {
    value: json['value'] == null ? undefined : json['value'],
    type: json['type'] == null ? undefined : V2beta1IOTypeFromJSON(json['type']),
    parameter_key: json['parameter_key'] == null ? undefined : json['parameter_key'],
    producer: json['producer'] == null ? undefined : V2beta1IOProducerFromJSON(json['producer']),
  };
}

export function InputOutputsIOParameterToJSON(json: any): InputOutputsIOParameter {
  return InputOutputsIOParameterToJSONTyped(json, false);
}

export function InputOutputsIOParameterToJSONTyped(
  value?: InputOutputsIOParameter | null,
  ignoreDiscriminator: boolean = false,
): any {
  if (value == null) {
    return value;
  }

  return {
    value: value['value'],
    type: V2beta1IOTypeToJSON(value['type']),
    parameter_key: value['parameter_key'],
    producer: V2beta1IOProducerToJSON(value['producer']),
  };
}