// frontend/src/apisv2beta1/run/models/PipelineTaskTaskState.ts
/* tslint:disable */
/* eslint-disable */
/**
 * backend/api/v2beta1/run.proto
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: version not set
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * - RUNTIME_STATE_UNSPECIFIED: Default value. This value is not used.
 *  - RUNNING: Entity execution is in progress.
 *  - SUCCEEDED: Entity completed successfully.
 *  - SKIPPED: Entity has been skipped. For example, due to caching.
 *  - FAILED: Entity execution has failed.
 * @export
 */
export const PipelineTaskTaskState = {
  RUNTIME_STATE_UNSPECIFIED: 'RUNTIME_STATE_UNSPECIFIED',
  RUNNING: 'RUNNING',
  SUCCEEDED: 'SUCCEEDED',
  SKIPPED: 'SKIPPED',
  FAILED: 'FAILED',
  CACHED: 'CACHED',
} as const;
export type PipelineTaskTaskState =
  (typeof PipelineTaskTaskState)[keyof typeof PipelineTaskTaskState];

export function instanceOfPipelineTaskTaskState(value: any): boolean {
  for (const key in PipelineTaskTaskState) {
    if (Object.prototype.hasOwnProperty.call(PipelineTaskTaskState, key)) {
      if (PipelineTaskTaskState[key as keyof typeof PipelineTaskTaskState] === value) {
        return true;
      }
    }
  }
  return false;
}

export function PipelineTaskTaskStateFromJSON(json: any): PipelineTaskTaskState {
  return PipelineTaskTaskStateFromJSONTyped(json, false);
}

export function PipelineTaskTaskStateFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): PipelineTaskTaskState {
  return json as PipelineTaskTaskState;
}

export function PipelineTaskTaskStateToJSON(value?: PipelineTaskTaskState | null): any {
  return value as any;
}

export function PipelineTaskTaskStateToJSONTyped(
  value: any,
  ignoreDiscriminator: boolean,
): PipelineTaskTaskState {
  return value as PipelineTaskTaskState;
}