// frontend/src/apisv2beta1/run/models/V2beta1Artifact.ts
export interface V2beta1Artifact {
  /**
   *
   * @type {string}
   * @memberof V2beta1Artifact
   */
  readonly artifact_id?: string;
  /**
   * Required. The client provided name of the artifact.
   * Note: in MLMD when name was set, it had to be unique for that type_id
   * this restriction is removed here
   * If this is a "Metric" artifact, the name of the metric
   * is treated as the Key in its K/V pair.
   * @type {string}
   * @memberof V2beta1Artifact
   */
  name?: string;
  /**
   *
   * @type {string}
   * @memberof V2beta1Artifact
   */
  description?: string;
  /**
   *
   * @type {ArtifactArtifactType}
   * @memberof V2beta1Artifact
   */
  type?: ArtifactArtifactType;
  /**
   * The uniform resource identifier of the physical artifact.
   * May be empty if there is no physical artifact.
   * @type {string}
   * @memberof V2beta1Artifact
   */
  uri?: string;
  /**
   * Optional. User provided custom properties which are not defined by its type.
   * @type {{ [key: string]: object; }}
   * @memberof V2beta1Artifact
   */
  metadata?: { [key: string]: object };
  /**
   *
   * @type {number}
   * @memberof V2beta1Artifact
   */
  number_value?: number;
  /**
   * Output only. Create time of the artifact in millisecond since epoch.
   * Note: The type and name is updated from mlmd artifact to be consistent with other backend apis.
   * @type {Date}
   * @memberof V2beta1Artifact
   */
  readonly created_at?: Date;
  /**
   *
   * @type {string}
   * @memberof V2beta1Artifact
   */
  namespace?: string;
}